+++ lib/humps.dart
@@ -1,11 +1,4 @@
-/// # Humps
-/// ## Features
-/// * toCamelCase() method to convert map keys from snake_case to camelCase
-/// * toSnakeCase() method to convert map keys from camelCase to snake_case
-/// * Works recursively on nested maps
-/// * Handles preserving leading underscores
+/// A Dart library for converting case style on strings and maps.
 library;
 
 export 'src/humps.dart';
-
-// TODO: Export any libraries intended for clients of this package.