+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.2
+
+- adding the `decamelize` method on `String` class to convert a camelCase string to snake_case or kebab-case
+- adding the `decamelizeKeys` method on `Map` class to convert the keys of a map from camelCase to snake_case or kebab-case
+- adding the `camelize` method on `String` class to convert a snake_case or kebab-case string to camelCase
+- adding the `camelizeKeys` method on `Map` class to convert the keys of a map from snake_case or kebab-case to camelCase
+- adding the `pascalize` method on `String` class to convert a snake_case or kebab-case string to PascalCase
+- adding the `pascalizeKeys` method on `Map` class to convert the keys of a map from snake_case or kebab-case to PascalCase
+- adding the `depascalize` method on `String` class to convert a PascalCase string to snake_case or kebab-case
+- adding the `depascalizeKeys` method on `Map` class to convert the keys of a map from PascalCase to snake_case or kebab-case
+
 ## 1.0.1
 
 - Updating the docs